change hard coded addresses in program, some style tweaks to boot sector
the program has some hard coded addresses that assume certain parts of memory
are used for different things, particularly the address of the _ax variable
that is used for an interface to asm. the source code and machine code areas
of ram were swapped, so the program had to update.

also move several note related documents into a subfolder